requirements.txt
alabaster==0.7.12
argh==0.26.2
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.3
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dlib @ https://files.pythonhosted.org/packages/0e/ce/f8a3cff33ac03a8219768f0694c5d703c8e037e6aba2e865f9bae22ed63c/dlib-19.8.1-cp36-cp36m-win_amd64.whl#sha256=794994fa2c54e7776659fddb148363a5556468a6d5d46be8dad311722d54bfcf
docutils==0.16
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
flake8==3.7.9
future==0.18.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imagesize==1.2.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593431910144/work
intervaltree==3.0.2
ipykernel==5.3.2
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594163838128/work
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring @ file:///C:/ci/keyring_1593109797081/work
kiwisolver==1.2.0
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
opencv-contrib-python==4.3.0.36
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1592848020240/work
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
PyNaCl @ file:///C:/ci/pynacl_1593105438052/work
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rope==0.17.0
Rtree==0.9.4
scikit-learn==0.23.1
scipy==1.5.1
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx @ file:///tmp/build/80754af9/sphinx_1594223420021/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
watchdog @ file:///C:/ci/watchdog_1593446204230/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zipp==3.1.0
